refactor completed. organized files into appropraite folders
