+++ lib/poker_hand_eval_helpers.rb
