+++ lib/poker_hand_eval_libraries.rb
